requirements.txt
--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu118/
paddlepaddle-gpu==3.1.1
acres==0.5.0
aistudio-sdk==0.3.6
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.10.0
bce-python-sdk==0.9.45
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
colorlog==6.9.0
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
doclayout_yolo==0.0.4
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.19.1
fitz==0.0.1.dev2
fonttools==4.59.2
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
imagesize==1.4.1
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.6
kiwisolver==1.4.9
looseversion==1.3.0
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
modelscope==1.29.2
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==8.9.6.50
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle-gpu==3.1.1
paddlex==3.2.1
pandas==2.3.2
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.3.0
prettytable==3.16.0
protobuf==6.32.0
prov==2.1.1
psutil==7.0.0
puremagic==1.30
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pyparsing==3.2.3
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.1
simplejson==3.20.1
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
stringzilla==3.12.6
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
traits==7.0.2
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.13
